prototype/requirements.txt
adafruit-circuitpython-scd30
adafruit-circuitpython-sgp30
adafruit-circuitpython-sht31d
adafruit-circuitpython-sht4x
adafruit-circuitpython-veml7700
adafruit-circuitpython-tca9548a
python3-gpiozero


--pip3 install--